bebig-136 refactor : UserService 인터페이스 생성, 트랜젝션 관리
